requirements/base.txt
Django==1.11.7
psycopg2
django-background-tasks
Pillow
sorl-thumbnail
